added sass functionality Sass will resolve our conflicting css by allowing us to nest styles and create unique css rules per unique component.
