+++ frontend/src/admin/components/accounts/AdminAccounts.js
@@ -1,17 +1,16 @@
 import React from 'react'
 import {Link} from 'react-router-dom'
-import AdminHeader from './adminHeader/adminHeader.jsx'
+import AdminHeader from '../adminHeader/adminHeader.jsx'
 
 
-const AdminData = () => {
+const AdminAccounts = () => {
 return (
     <div>
-    <AdminHeader selected='Data' username='Admin T.'></AdminHeader>
+    <AdminHeader selected='Accounts' username='Admin T.'></AdminHeader>
         <div className='AdminContentArea'>
         !!!code goes here!!!
         </div>
     </div>
-   
 )};
 
-export default AdminData 
+export default AdminAccounts
