+++ frontend/src/admin/components/adminEvent/adminEvent.jsx
@@ -1,7 +1,7 @@
 import React, {useState} from 'react'
 import {Link} from 'react-router-dom'
 
-import '../../styleSheets/Admin-events-style.css'
+import '../events/adminEvents.css'
 import dropDownIcon from '../../icons/arrow_drop_down-24px.svg';
 import editIcon from '../../icons/edit-24px.svg';
