+++ frontend/src/admin/components/adminHeader/adminHeader.css
@@ -1,212 +1,218 @@
-.Banner {
-    display: flex;
-    flex-direction: row;
-    background: #C4C4C4;
-    height: 50px;
+.adminHeader .Banner {
+  display: -webkit-box;
+  display: -ms-flexbox;
+  display: flex;
+  -webkit-box-orient: horizontal;
+  -webkit-box-direction: normal;
+      -ms-flex-direction: row;
+          flex-direction: row;
+  background: #C4C4C4;
+  height: 50px;
 }
 
-.bannerSpacer{
-    margin-right:auto;
+.adminHeader .bannerSpacer {
+  margin-right: auto;
 }
 
-.settingIcon {
-    width: 24px;
-    margin-right: 15px;
-    /* height: 24px; */
+.adminHeader .settingIcon {
+  width: 24px;
+  margin-right: 15px;
+  /* height: 24px; */
 }
 
-.quickTools {
-    display: flex;
-    flex-direction: row;
-    background: #F3F3F3;
-    height: 40px;
+.adminHeader .quickTools {
+  display: -webkit-box;
+  display: -ms-flexbox;
+  display: flex;
+  -webkit-box-orient: horizontal;
+  -webkit-box-direction: normal;
+      -ms-flex-direction: row;
+          flex-direction: row;
+  background: #F3F3F3;
+  height: 40px;
 }
 
-.debug {
-    border: solid red 1px;
+.adminHeader .debug {
+  border: solid red 1px;
 }
 
-.quickToolSeperator {
-    margin-left: 0px;
-    margin-right: 0px;
+.adminHeader .quickToolSeperator {
+  margin-left: 0px;
+  margin-right: 0px;
 }
 
-.quickToolSpacer {
-    width: 0px;
-    margin-right: auto;
+.adminHeader .quickToolSpacer {
+  width: 0px;
+  margin-right: auto;
 }
 
-.tab {
-    margin: 0px;
+.adminHeader .tab {
+  margin: 0px;
 }
 
-.quickToolName {
-    display: inline-block;
-    position: relative;
-    bottom: 5px;
-    padding-left: 8px;
-    padding-right: 8px;
+.adminHeader .quickToolName {
+  display: inline-block;
+  position: relative;
+  bottom: 5px;
+  padding-left: 8px;
+  padding-right: 8px;
 }
 
-.plusResize {
-    position: relative;
-    right: 10px;
-    top: 5px;
-    height: 15px;
-    background-color: #ffffff;
-    border-radius: 15px;
+.adminHeader .plusResize {
+  position: relative;
+  right: 10px;
+  top: 5px;
+  height: 15px;
+  background-color: #ffffff;
+  border-radius: 15px;
 }
 
-.centered {
-    margin-top: 6px;
-    /* margin-left: 7px; */
+.adminHeader .centered {
+  margin-top: 6px;
+  /* margin-left: 7px; */
 }
 
-.leftMargin {
-    margin-left: 7px;
+.adminHeader .leftMargin {
+  margin-left: 7px;
 }
 
-.selected {
-    position: relative;
-    height: 100%;
-    background: white;
-    border-top-left-radius: 10px;
-    border-top-right-radius: 10px;
-    /* border:solid white 2px; */
-    box-shadow:0px -3px 3px rgba(0, 0, 0, 0.15);
+.adminHeader .selected {
+  position: relative;
+  height: 100%;
+  background: white;
+  border-top-left-radius: 10px;
+  border-top-right-radius: 10px;
+  /* border:solid white 2px; */
+  -webkit-box-shadow: 0px -3px 3px rgba(0, 0, 0, 0.15);
+          box-shadow: 0px -3px 3px rgba(0, 0, 0, 0.15);
 }
 
-.logoResize {
-    margin: 5px;
-    height: 40px;
+.adminHeader .logoResize {
+  margin: 5px;
+  height: 40px;
 }
 
-.username {
-    position: relative;
-    font-family: 'Roboto', sans-serif;
-    margin-right: 20px;
-    top: 15px;
+.adminHeader .username {
+  position: relative;
+  font-family: 'Roboto', sans-serif;
+  margin-right: 20px;
+  top: 15px;
 }
 
-.rotate90 {
-    transform: rotate(-90deg);
+.adminHeader .rotate90 {
+  -webkit-transform: rotate(-90deg);
+          transform: rotate(-90deg);
 }
 
-.dropDownIcon {
-    position: relative;
-    top: 13px;
-    height: 24px;
+.adminHeader .dropDownIcon {
+  position: relative;
+  top: 13px;
+  height: 24px;
 }
 
-.settingsMenu {
-    position: absolute;
-    top: 45px;
-    left: calc(100% - 260px);
-    z-index: 2;
-    padding-bottom: 10px;
-    width: 240px;
-    background: white;
-    box-shadow: 0px 4px 10px rgba(0, 0, 0, 0.25);
-    border-radius: 5px;
+.adminHeader .settingsMenu {
+  position: absolute;
+  top: 45px;
+  left: calc(100% - 260px);
+  z-index: 2;
+  padding-bottom: 10px;
+  width: 240px;
+  background: white;
+  -webkit-box-shadow: 0px 4px 10px rgba(0, 0, 0, 0.25);
+          box-shadow: 0px 4px 10px rgba(0, 0, 0, 0.25);
+  border-radius: 5px;
 }
 
-.settingsOption {
-    margin: 0px;
-    padding-left: 10px;
-    color: black;
-    
+.adminHeader .settingsOption {
+  margin: 0px;
+  padding-left: 10px;
+  color: black;
 }
 
-.topOption {
-    padding-top: 10px;
+.adminHeader .topOption {
+  padding-top: 10px;
 }
 
-.hide {
-    z-index: -2;
+.adminHeader .hide {
+  z-index: -2;
 }
 
-.none {
-   padding: 0px;
-   margin: 0;
+.adminHeader .none {
+  padding: 0px;
+  margin: 0;
 }
 
-#hidden {
-    visibility: hidden;
+.adminHeader #hidden {
+  visibility: hidden;
 }
 
-
-#createEventFix {
-    right: 10px;
-    padding-right: 0px;
+.adminHeader #createEventFix {
+  right: 10px;
+  padding-right: 0px;
 }
 
-#endSpacer {
-    width: 0px;
-    margin-left: 15px;
+.adminHeader #endSpacer {
+  width: 0px;
+  margin-left: 15px;
 }
 
-
-
-.settingsOption button {
-    display: inline-block;
-    border: none;
-    width: calc(100% - 10px);
-    padding: 0px;
-    margin-right: 20px;
-    text-decoration: none;
-    background: #ffffff;
-    color: #000000;
-    font-family: sans-serif;
-    font-size: 1rem;
-    cursor: pointer;
-    text-align: left;
-    transition: background 250ms ease-in-out, 
-                transform 150ms ease;
-    -webkit-appearance: none;
-    -moz-appearance: none;
+.adminHeader .settingsOption button {
+  display: inline-block;
+  border: none;
+  width: calc(100% - 10px);
+  padding: 0px;
+  margin-right: 20px;
+  text-decoration: none;
+  background: #ffffff;
+  color: #000000;
+  font-family: sans-serif;
+  font-size: 1rem;
+  cursor: pointer;
+  text-align: left;
+  -webkit-transition: background 250ms ease-in-out,  -webkit-transform 150ms ease;
+  transition: background 250ms ease-in-out,  -webkit-transform 150ms ease;
+  transition: background 250ms ease-in-out,  transform 150ms ease;
+  transition: background 250ms ease-in-out,  transform 150ms ease,  -webkit-transform 150ms ease;
+  -webkit-appearance: none;
+  -moz-appearance: none;
 }
 
-
-.settingsOption button:hover,
-.settingsOption button:focus {
-    background: lightblue;
+.adminHeader .settingsOption button:hover,
+.adminHeader .settingsOption button:focus {
+  background: lightblue;
 }
 
-.settingsOption button:focus {
-    
-    outline-offset: -4px;
+.adminHeader .settingsOption button:focus {
+  outline-offset: -4px;
 }
 
-.settingsOption button:active {
-    transform: scale(0.99);
+.adminHeader .settingsOption button:active {
+  -webkit-transform: scale(0.99);
+          transform: scale(0.99);
 }
 
-li {
-    list-style:none;
+.adminHeader li {
+  list-style: none;
 }
 
 @media only screen and (max-width: 700px) {
-    .quickToolName{
-        visibility: hidden;
-        width: 0;
-    }
-
-   #mobileFix_1{
-       position: relative;
-        bottom: 8px;
-    }
-
-    .plusResize {
-        top: -4px;
-        right: 4px;
-    }
-
-    .centered {
-        margin-top: 6px;
-        position: relative;
-        left: 4px;
-
-        /* margin-left: 7px; */
-    }
-    
+  .adminHeader .quickToolName {
+    visibility: hidden;
+    width: 0;
   }
+  .adminHeader #mobileFix_1 {
+    position: relative;
+    bottom: 8px;
+  }
+  .adminHeader .plusResize {
+    top: -4px;
+    right: 4px;
+  }
+  .adminHeader .centered {
+    margin-top: 6px;
+    position: relative;
+    left: 4px;
+    /* margin-left: 7px; */
+  }
+}
+/*# sourceMappingURL=adminHeader.css.map */
