+++ frontend/src/admin/components/login/adminLogin.css
@@ -1,0 +1,94 @@
+.adminLogin {
+  /*h2{
+    font-family: 'Roboto', sans-serif;
+    font-style: normal;
+    font-weight: normal;
+    font-size: 64px;
+    color: black;
+    line-height: 0px;
+}*/
+}
+
+.adminLogin .log-in-container {
+  display: -webkit-box;
+  display: -ms-flexbox;
+  display: flex;
+  -webkit-box-align: center;
+      -ms-flex-align: center;
+          align-items: center;
+  -webkit-box-pack: center;
+      -ms-flex-pack: center;
+          justify-content: center;
+  text-align: center;
+  height: 100%;
+  -webkit-box-orient: horizontal;
+  -webkit-box-direction: normal;
+      -ms-flex-direction: row;
+          flex-direction: row;
+}
+
+.adminLogin .log-in-background {
+  border-radius: 50px;
+  height: 80%;
+  padding: 20px;
+  width: 500px;
+  height: 450px;
+}
+
+.adminLogin .top-banner {
+  display: -webkit-box;
+  display: -ms-flexbox;
+  display: flex;
+  width: 100vw;
+  background-color: #C4C4C4;
+  height: 15vh;
+}
+
+.adminLogin .image-size {
+  /* border: 1px solid; */
+  width: 18%;
+  margin-top: 17px;
+  /* width: 25%;
+    margin-left: -365px;
+    margin-top: 32px; */
+}
+
+.adminLogin .signin-form-2 {
+  padding-top: 47px;
+}
+
+.adminLogin .signin-form-2 input[type='text'] {
+  margin-top: 22px;
+  font-size: 20px;
+  padding-left: 10px;
+  border-radius: 11px;
+  background-color: #CFC8C8;
+  color: #FFFFFF;
+  width: 30vw;
+  height: 48px;
+}
+
+.adminLogin .signin-form-2 input::-webkit-input-placeholder {
+  font-size: 20px;
+  color: #FFFFFF;
+}
+
+.adminLogin .signin-form-2 input:-ms-input-placeholder {
+  font-size: 20px;
+  color: #FFFFFF;
+}
+
+.adminLogin .signin-form-2 input::-ms-input-placeholder {
+  font-size: 20px;
+  color: #FFFFFF;
+}
+
+.adminLogin .signin-form-2 input::placeholder {
+  font-size: 20px;
+  color: #FFFFFF;
+}
+
+.adminLogin .signin-form-2 input[type='submit'] {
+  margin-top: 37px;
+}
+/*# sourceMappingURL=adminLogin.css.map */
