+++ frontend/src/admin/components/login/adminLogin.scss
@@ -1,5 +1,6 @@
+.adminLogin {
 
-/*h2{
+    /*h2{
     font-family: 'Roboto', sans-serif;
     font-style: normal;
     font-weight: normal;
@@ -65,3 +66,5 @@
   .signin-form-2 input[type='submit'] {
     margin-top: 37px;
   }
+
+}
